// src/app/api_codegen/model/createScheduleDto.ts
/**
 * Invicto API
 * REST API Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AssignHolidayDto } from './assignHolidayDto';
import { CreateTimetableDto } from './createTimetableDto';

export interface CreateScheduleDto { 
    name: string;
    description?: string;
    timetables: Array<CreateTimetableDto>;
    holidays?: Array<AssignHolidayDto>;
}